Update BookingPage.js
added test cases

Added a UserProfile page

Users can see their bookings, taps to the database to see the number of tickets booked and for what event.

Added a Booking success page and a user profile page and implement backend logic to record user bookings based on userID and update number of available tickets

Added a Booking page to selected number of tickets and success page

Enables user to select desired number of tickets and updates the number of available tickets in the database.